Split tests by type
